+++ apps/web/src/services/dashboards/constants/managed-variables-schema.ts
@@ -1,6 +1,7 @@
+import type { DashboardVariablesSchema } from '@/schema/dashboard/_types/dashboard-type';
+
 import { MANAGED_VARIABLE_MODEL_CONFIGS } from '@/lib/variable-models/managed';
 
-import type { DashboardVariablesSchema } from '@/services/dashboards/types/dashboard-model-type';
 
 export const MANAGED_DASH_VAR_SCHEMA: DashboardVariablesSchema = {
     properties: {
